backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
scipy==1.12.0
joblib==1.3.2
torch==2.2.1
transformers==4.39.3
requests==2.31.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
tqdm==4.66.2  # For progress tracking
protobuf==4.25.3  # Avoids compatibility issues
sentencepiece==0.1.99  # Optional, for tokenizer support